requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.4.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
ca-certificates=2024.2.2=hbcca054_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
comm=0.2.1=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.12.1=pypi_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
filelock=3.13.1=pypi_0
fonttools=4.49.0=pypi_0
gdown=5.1.0=pypi_0
idna=3.4=py39h06a4308_0
importlib-resources=6.1.3=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.3=pyhd33586a_0
ipython=8.12.0=pyh41d4057_0
jedi=0.19.1=pyhd8ed1ab_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.1=py39hf3d152e_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libuv=1.44.2=h5eee18b_0
lxml=5.1.0=pypi_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h213fc3f_46344
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.26.4=pypi_0
openssl=3.0.13=h7f8727e_0
packaging=23.2=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=23.3.1=py39h06a4308_0
platformdirs=4.2.0=pyhd8ed1ab_0
portalocker=2.8.2=pypi_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=pypi_0
python=3.9.18=h955ad1f_0
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.9.0=py3.9_cuda10.2_cudnn7.6.5_0
pyzmq=25.1.2=py39h6a678d5_0
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests=2.31.0=py39h06a4308_1
sacrebleu=2.4.0=pypi_0
setuptools=68.2.2=py39h06a4308_0
six=1.16.0=pyh6c4a22f_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
torchtext=0.10.0=py39
tornado=6.1=py39hb9d737c_3
tqdm=4.65.0=py39hb070fc8_0
traitlets=5.14.1=pyhd8ed1ab_0
typing_extensions=4.9.0=py39h06a4308_1
tzdata=2024a=h04d1e81_0
urllib3=2.1.0=py39h06a4308_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.41.2=py39h06a4308_0
xz=5.4.6=h5eee18b_0
zeromq=4.3.5=h6a678d5_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
